add tests and fix imports #222
